42-FE: `PageLayout` Add `isNarrow` prop & move to different folder
