--- Okane.Client/src/shared/layouts/PageLayout.vue
@@ -1,19 +1,0 @@
-<template>
-  <div class="root">
-    <slot />
-  </div>
-</template>
-
-<style scoped lang="scss">
-.root {
-  margin-left: var(--space-md);
-  margin-right: var(--space-md);
-  max-width: 50em;
-}
-
-@include respond(md) {
-  .root {
-    margin: 0 auto;
-  }
-}
-</style>
