+++ Okane.Client/src/shared/pages/AccountDeletedPage.vue
@@ -1,13 +1,13 @@
 <script setup lang="ts">
 // Internal
 import Heading from '@shared/components/Heading.vue'
-import PageLayout from '@shared/layouts/PageLayout.vue'
+import PageLayout from '@shared/components/wrappers/PageLayout.vue'
 
 import { AUTH_COPY } from '@features/auth/constants/copy'
 </script>
 
 <template>
-  <PageLayout>
+  <PageLayout is-narrow>
     <Heading tag="h1">{{ AUTH_COPY.ACCOUNT_DELETED.HEADING }}</Heading>
     <p>{{ AUTH_COPY.ACCOUNT_DELETED.TEXT }}</p>
   </PageLayout>
