+++ Okane.Client/src/shared/pages/ResetPasswordPage.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 // Internal
-import PageLayout from '@shared/layouts/PageLayout.vue'
+import PageLayout from '@shared/components/wrappers/PageLayout.vue'
 import ResetPasswordForm from '@features/auth/components/ResetPasswordForm.vue'
 
 import { useQueryPasswordRequirements } from '@features/auth/composables/useQueryPasswordRequirements'
@@ -9,7 +9,7 @@ const { data: requirements } = useQueryPasswordRequirements()
 </script>
 
 <template>
-  <PageLayout>
+  <PageLayout is-narrow>
     <ResetPasswordForm v-if="requirements" :requirements="requirements" />
   </PageLayout>
 </template>
